remove files works

+++ lib/zipper.rb
@@ -2,7 +2,7 @@ class Zipper
   
   class << self  
 
-    def unzip(zip_path)
+    def unzip(zip_path, delete: true)
       raise NoBlockGiven, "You need to pass a block with argument unzipped_path" unless block_given?
 
       basename = File.basename(zip_path, ".*")  # remove extension too
@@ -12,10 +12,10 @@ class Zipper
 
       yield unzipped_path
 
-      `rm -rf #{unzipped_path}`
+      `rm -rf #{unzipped_path}` if delete
     end
 
-    def zip(path)
+    def zip(path, delete: true)
       raise NoBlockGiven, "You need to pass a block with argument zipped_path" unless block_given?
       raise NotADirectory unless File.directory?(path)
 
@@ -28,7 +28,17 @@ class Zipper
 
       yield zipped_path
 
-      `rm #{zipped_path}`
+      `rm #{zipped_path}` if delete
+    end
+
+    def remove_multimedia_files(directory)
+      files = `find #{directory} -type f -and -print0`.split("\0").map{ |f| Shellwords.escape(f)}
+      files.each do |file|
+        mime_type = `file --brief --mime #{file}`
+        `rm #{file}` if ["image", "video", "audio"].any? { |type| mime_type.include?(type) }
+      end
+
+      true
     end
 
   end
